Refactor for slightly improved isolation. #41
